Am facut modificari
